requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Music APIs
spotipy>=2.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking (optional but recommended)
wandb>=0.15.0
tensorboard>=2.13.0

# Hyperparameter Optimization
optuna>=3.2.0

# Interpretability
shap>=0.42.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.23.0
